creating common dropbox manager
